backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.0"}
pydantic = ">=2.7"
langchain-openai = "^0.2.2"
python-decouple = "^3.8"
langchain-ollama = "^0.2.0"
langchain = "^0.3.3"
langchain-qdrant = "^0.1.4"
langchain-redis = "^0.1.0"
redis = "^5.1.1"
qdrant-client = "^1.12.0"
redisvl = "0.3.3"
langsmith = "^0.1.135"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
